refactor(session): split session namespace responsibilities
