+++ packages/session/AGENTS.md
@@ -9,7 +9,10 @@ src/
 ├── index.ts              # Package barrel — re-exports all namespaces
 ├── bus/                  # Bus pub/sub (Bus.publish / Bus.subscribe) + typed event descriptors
 ├── session/
-│   ├── index.ts          # Session namespace: CRUD, messages/parts, child sessions, worker meta
+│   ├── index.ts          # Session namespace barrel: public Session.* API re-exports
+│   ├── events.ts         # Session bus event descriptors
+│   ├── lifecycle.ts      # Session CRUD, child sessions, worker meta, TTL lazy deletion
+│   ├── messages.ts       # Message/part writes, pagination, hydration, resume recovery
 │   └── info.ts           # SessionInfo schema (leaf — breaks session ↔ storage cycle)
 ├── storage/
 │   ├── index.ts          # Barrel
